rename component to PasswordDoubleInput
