+++ tchap/resources/templates/pages/register/password.html
@@ -10,9 +10,6 @@ Please see LICENSE files in the repository root for full details.
 
 {% block content %}
 
-{{ include_asset('src/external/View.tsx') | indent(4) | safe }}
-
-
   <header class="page-heading">
     <div class="icon">
       {{ icon.user_profile_solid() }}
@@ -46,6 +43,10 @@ Please see LICENSE files in the repository root for full details.
     {% endcall %}    
     <!-- :tchap: end -->
     
+    <!-- use the react component for password creation with complexity and hints -->
+    {{ include_asset('src/external/register/PasswordDoubleInput.tsx') | indent(4) | safe }}
+    <div id="password-double-input" data-props='{"forceShowNewPasswordInvalid": true}'></div>
+
     <noscript>
       <!-- use the static password fields if javascript is not available in browser -->
       {% call(f) field.field(label=_("common.password"), name="password", form_state=form) %}
@@ -56,13 +57,6 @@ Please see LICENSE files in the repository root for full details.
         <input {{ field.attributes(f) }} class="cpd-text-control" type="password" autocomplete="new-password" required />
       {% endcall %}
     </noscript>
-
-    <div id="view" data-component="PasswordFormLoader" data-props='{"forceShowNewPasswordInvalid": false}'></div>
-    
-    <!-- <script type="module" src="/assets/View.js"></script> -->
-
-
-
     {{ button.button(text=_("action.continue")) }}
 
     {% set params = next["params"] | default({}) | to_params(prefix="?") %}
