+++ tchap/resources/translations/en.json
@@ -10,7 +10,7 @@
     },
     "continue": "Continue",
     "@continue": {
-      "context": "form_post.html:25:28-48, pages/consent.html:74:28-48, pages/device_consent.html:124:13-33, pages/device_link.html:40:26-46, pages/login.html:112:30-50, pages/reauth.html:32:28-48, pages/recovery/start.html:38:26-46, pages/register/password.html:66:26-46, pages/register/steps/display_name.html:45:28-48, pages/register/steps/registration_token.html:41:28-48, pages/register/steps/verify_email.html:51:26-46, pages/sso.html:37:28-48"
+      "context": "form_post.html:25:28-48, pages/consent.html:74:28-48, pages/device_consent.html:124:13-33, pages/device_link.html:40:26-46, pages/login.html:112:30-50, pages/reauth.html:32:28-48, pages/recovery/start.html:38:26-46, pages/register/password.html:60:26-46, pages/register/steps/display_name.html:45:28-48, pages/register/steps/registration_token.html:41:28-48, pages/register/steps/verify_email.html:51:26-46, pages/sso.html:37:28-48"
     },
     "create_account": "Create Account",
     "@create_account": {
@@ -79,7 +79,7 @@
     },
     "email_address": "Email address",
     "@email_address": {
-      "context": "pages/recovery/start.html:34:33-58, pages/register/password.html:44:33-58, pages/upstream_oauth2/do_register.html:120:37-62"
+      "context": "pages/recovery/start.html:34:33-58, pages/register/password.html:41:33-58, pages/upstream_oauth2/do_register.html:120:37-62"
     },
     "hide": "",
     "@hide": {},
@@ -93,17 +93,17 @@
     },
     "password": "Password",
     "@password": {
-      "context": "pages/login.html:100:37-57, pages/reauth.html:28:35-55, pages/register/password.html:51:35-55"
+      "context": "pages/login.html:100:37-57, pages/reauth.html:28:35-55, pages/register/password.html:52:35-55"
     },
     "password_confirm": "Confirm password",
     "@password_confirm": {
-      "context": "pages/register/password.html:55:35-63"
+      "context": "pages/register/password.html:56:35-63"
     },
     "show": "",
     "@show": {},
     "username": "Username",
     "@username": {
-      "context": "pages/login.html:94:37-57, pages/register/index.html:59:35-55, pages/register/password.html:40:45-65, pages/upstream_oauth2/do_register.html:104:35-55, pages/upstream_oauth2/do_register.html:109:39-59"
+      "context": "pages/login.html:94:37-57, pages/register/index.html:59:35-55, pages/register/password.html:37:45-65, pages/upstream_oauth2/do_register.html:104:35-55, pages/upstream_oauth2/do_register.html:109:39-59"
     }
   },
   "create_account": "",
@@ -619,7 +619,7 @@
     "register": {
       "call_to_login": "Already have an account?",
       "@call_to_login": {
-        "context": "pages/register/index.html:89:35-66, pages/register/password.html:69:33-64",
+        "context": "pages/register/index.html:89:35-66, pages/register/password.html:63:33-64",
         "description": "Displayed on the registration page to suggest to log in instead"
       },
       "continue_with_email": "Continue with email address",
@@ -633,7 +633,7 @@
         },
         "heading": "Create an account",
         "@heading": {
-          "context": "pages/register/password.html:22:27-67"
+          "context": "pages/register/password.html:19:27-67"
         }
       },
       "terms_of_service": "I agree to the <a href=\"%s\" data-kind=\"primary\" class=\"cpd-link\">Terms and Conditions</a>",
